+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add ThreeDSecure test payment method nonces
 * Add test `AuthenticationId`s
 * Add `DisputeAccepted`, `DisputeDisputed`, and `DisputeExpired` webhook constants
+* Add `Authentication Insight` to payment method nonce create
 
 # 2.101.0
 * Add `RefundAuthHardDeclined` and `RefundAuthSoftDeclined` to validation errors